refactor: `Memo`와 `MemoHeader`를 분리한다
