+++ packages/shared/src/hooks/useTabQuery.ts
@@ -1,7 +1,7 @@
 import { queryKeys } from '@src/constants';
 import { Tab } from '@src/utils/extension';
-import { useQuery } from '@tanstack/react-query';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function useTabQuery() {
-  return useQuery({ queryFn: Tab.get, queryKey: queryKeys.tab() });
+  return useSuspenseQuery({ queryFn: Tab.get, queryKey: queryKeys.tab() });
 }
