+++ pages/side-panel/src/components/index.ts
@@ -3,4 +3,5 @@ export { default as Memo } from './Memo';
 export { default as Header } from './Header';
 export { default as SummaryProvider } from './SummaryProvider';
 export { default as LoginSection } from './LoginSection';
+export { default as MemoHeader } from './MemoHeader';
 export * from './SummaryProvider';
